Move module-level `PackageManager` definitions back to `PackageManager.py`
